Add modules and make the classes functionable
