+++ index.html
@@ -72,6 +72,6 @@
     <p>Copyright</p>
 </footer>
 
-<script src="index.js"></script>
+<script src="index.js" type="module"></script>
 </body>
 </html>
